implemented new app routing and route guard
