+++ frontend/src/components/layout/TopBar.vue
@@ -1,0 +1,16 @@
+
+<template lang="">
+    <div>
+        
+    </div>
+</template>
+
+<script lang="ts">
+    export default {
+        
+    }
+</script>
+
+<style lang="">
+    
+</style>
