+++ frontend/src/modules/app/views/admin/router/index.ts
@@ -1,0 +1,15 @@
+export default [
+    {
+        path: '/app/admin',
+        name: 'admin',
+        redirect: { name: 'adminMain' },
+        component: () => import(/* webpackChunkName: "adminMain" */ '@/modules/app/views/admin/views/Main.vue'),
+        children: [
+          {
+            path: '/main',
+            name: 'adminMain',
+            component: () => import(/* webpackChunkName: "adminMain" */ '@/modules/app/views/admin/views/Main.vue')
+          },
+        ]
+    },
+]
