+++ frontend/src/modules/app/views/dashboard/router/index.ts
@@ -1,0 +1,15 @@
+export default [
+    {
+        path: '/app/dashboard',
+        name: 'dashboard',
+        redirect: { name: 'dashboardMain' },
+        component: () => import(/* webpackChunkName: "dashboardMain" */ '@/modules/app/views/dashboard/views/Main.vue'),
+        children: [
+          {
+            path: '/app/dashboard/main',
+            name: 'dashboardMain',
+            component: () => import(/* webpackChunkName: "dashboardMain" */ '@/modules/app/views/dashboard/views/Main.vue')
+          },
+        ]
+    },
+]
