+++ frontend/src/modules/app/views/dashboard/views/DashboardSideview.vue
