+++ frontend/src/modules/app/views/dashboard/views/Main.vue
