+++ frontend/src/modules/app/views/events/views/Main.vue
