+++ frontend/src/modules/app/views/groups/views/Main.vue
