+++ frontend/src/modules/app/views/management/views/Main.vue
