+++ frontend/src/modules/app/views/statistic/router/index.ts
@@ -1,0 +1,15 @@
+export default [
+    {
+        path: '/app/statistic',
+        name: 'statistic',
+        redirect: { name: 'statisticMain' },
+        component: () => import(/* webpackChunkName: "statisticMain" */ '@/modules/app/views/statistic/views/Main.vue'),
+        children: [
+          {
+            path: '/app/statistic/main',
+            name: 'statisticMain',
+            component: () => import(/* webpackChunkName: "statisticMain" */ '@/modules/app/views/statistic/views/Main.vue')
+          },
+        ]
+    },
+]
