+++ frontend/src/modules/app/views/statistic/views/Main.vue
