--- frontend/src/modules/event/router/index.ts
@@ -1,15 +1,0 @@
-export default [
-    {
-        path: '/event',
-        name: 'event',
-        redirect: { name: 'eventMain' },
-        component: () => import(/* webpackChunkName: "register" */ '@/modules/event/views/Main.vue'),
-        children: [
-          {
-            path: '/event/main',
-            name: 'eventMain',
-            component: () => import(/* webpackChunkName: "register" */ '@/modules/event/views/Main.vue')
-          },
-        ]
-    },
-]
