+++ frontend/src/modules/home/router/index.ts
@@ -1,0 +1,8 @@
+export default [
+    {
+        path: '/',
+        name: 'home',
+        alias: '/home',
+        component: () => import(/* webpackChunkName: "Home" */ '@/modules/home/views/Main.vue'),
+    },
+]
