+++ frontend/src/modules/home/views/Main.vue
@@ -1,0 +1,13 @@
+<template lang="">
+  <div>
+    Home
+  </div>
+</template>
+<script lang="ts">
+export default {
+  
+}
+</script>
+<style lang="">
+  
+</style>
