--- frontend/src/modules/management/router/index.ts
@@ -1,15 +1,0 @@
-export default [
-    {
-        path: '/management',
-        name: 'management',
-        redirect: { name: 'managementMain' },
-        component: () => import(/* webpackChunkName: "register" */ '@/modules/management/views/Main.vue'),
-        children: [
-          {
-            path: '/management/main',
-            name: 'managementMain',
-            component: () => import(/* webpackChunkName: "register" */ '@/modules/management/views/Main.vue')
-          },
-        ]
-    },
-]
