--- frontend/src/modules/member/router/index.ts
@@ -1,15 +1,0 @@
-export default [
-    {
-        path: '/member',
-        name: 'member',
-        redirect: { name: 'memberMain' },
-        component: () => import(/* webpackChunkName: "register" */ '@/modules/member/views/Main.vue'),
-        children: [
-          {
-            path: '/member/main',
-            name: 'memberMain',
-            component: () => import(/* webpackChunkName: "register" */ '@/modules/member/views/Main.vue')
-          },
-        ]
-    },
-]
